Refactor: Let the LLM service return the index manager
