+++ apps/assistants/tests/test_sync.py
@@ -11,7 +11,6 @@ from openai.pagination import SyncCursorPage
 from apps.assistants.models import ToolResources
 from apps.assistants.sync import (
     OpenAiSyncError,
-    OpenAIVectorStoreManager,
     _update_or_create_vector_store,
     delete_openai_assistant,
     get_out_of_sync_files,
@@ -354,7 +353,7 @@ class TestVectorStoreManager:
     @patch("openai.resources.vector_stores.file_batches.FileBatches.create")
     def test_link_files_to_vector_store(self, mock_file_batches_create):
         """Test linking files to vector store with and without chunking strategy"""
-        manager = OpenAIVectorStoreManager.from_llm_provider(LlmProviderFactory())
+        manager = LlmProviderFactory().get_index_manager()
         vector_store_id = "vs_123"
         file_ids = [f"file_{i}" for i in range(600)]  # Create more than 500 files to test batching
 
@@ -407,7 +406,7 @@ class TestVectorStoreManager:
     @patch("openai.resources.vector_stores.VectorStores.retrieve")
     def test_get(self, mock_retrieve):
         """Test retrieving a vector store"""
-        manager = OpenAIVectorStoreManager.from_llm_provider(LlmProviderFactory())
+        manager = LlmProviderFactory().get_index_manager()
         vector_store_id = "vs_123"
         expected_result = ObjectWithId(id=vector_store_id)
         mock_retrieve.return_value = expected_result
@@ -420,7 +419,7 @@ class TestVectorStoreManager:
     @patch("openai.resources.vector_stores.VectorStores.create")
     def test_create_vector_store(self, mock_create):
         """Test creating a vector store with and without files"""
-        manager = OpenAIVectorStoreManager.from_llm_provider(LlmProviderFactory())
+        manager = LlmProviderFactory().get_index_manager()
         expected_id = "vs_123"
         mock_create.return_value = ObjectWithId(id=expected_id)
 
@@ -440,7 +439,7 @@ class TestVectorStoreManager:
     @patch("openai.resources.vector_stores.VectorStores.delete")
     def test_delete_vector_store(self, mock_delete):
         """Test deleting a vector store with different error scenarios"""
-        manager = OpenAIVectorStoreManager.from_llm_provider(LlmProviderFactory())
+        manager = LlmProviderFactory().get_index_manager()
         vector_store_id = "vs_123"
 
         # Test successful deletion
@@ -463,7 +462,7 @@ class TestVectorStoreManager:
     @patch("openai.resources.vector_stores.files.Files.delete")
     def test_delete_file(self, mock_delete):
         """Test deleting a file from a vector store"""
-        manager = OpenAIVectorStoreManager.from_llm_provider(LlmProviderFactory())
+        manager = LlmProviderFactory().get_index_manager()
         vector_store_id = "vs_123"
         file_id = "file_123"
