+++ apps/documents/views.py
@@ -13,7 +13,7 @@ from django.views.decorators.http import require_POST
 from django.views.generic import CreateView, TemplateView, UpdateView
 from django_tables2 import SingleTableView
 
-from apps.assistants.sync import OpenAIVectorStoreManager, delete_file_from_openai
+from apps.assistants.sync import delete_file_from_openai
 from apps.documents.forms import CollectionForm
 from apps.documents.models import Collection, CollectionFile, FileStatus
 from apps.documents.tables import CollectionsTable
@@ -158,7 +158,7 @@ class CreateCollection(LoginAndTeamRequiredMixin, CollectionFormMixin, CreateVie
         response = super().form_valid(form)
         if form.instance.is_index:
             collection = form.instance
-            manager = OpenAIVectorStoreManager.from_llm_provider(collection.llm_provider)
+            manager = collection.llm_provider.get_index_manager()
             collection.openai_vector_store_id = manager.create_vector_store(name=collection.index_name)
             collection.save(update_fields=["openai_vector_store_id"])
 
@@ -189,7 +189,7 @@ class EditCollection(LoginAndTeamRequiredMixin, CollectionFormMixin, UpdateView)
 
         if form.instance.is_index and "llm_provider" in form.changed_data:
             with transaction.atomic():
-                new_manager = OpenAIVectorStoreManager.from_llm_provider(collection.llm_provider)
+                new_manager = collection.llm_provider.get_index_manager()
                 collection.openai_vector_store_id = new_manager.create_vector_store(collection.index_name)
                 collection.save(update_fields=["openai_vector_store_id"])
