+++ apps/service_providers/exceptions.py
@@ -2,3 +2,7 @@ class ServiceProviderConfigError(Exception):
     def __init__(self, provider_type: str, message: str):
         self.provider_type = provider_type
         super().__init__(f"[{provider_type}] provider config error: {message}")
+
+
+class OpenAiUnableToLinkFileError(Exception):
+    pass
